requirements.txt
# Core ML & CV Libraries
tensorflow>=2.13.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
pillow>=10.0.0

# Web Framework
flask>=2.3.0
flask-cors>=4.0.0

# Optional utilities
python-multipart>=0.0.6
